Move StartScreenGUI Screen into include/, src/ files
